Initial GitHub Commit
This is a project that Jacob has thrown together using most of everyone's code. None of it is working together yet, that's what our week's assignment is. And other than tweaking with the includes at the top of each file, I have not touched anyone's code other than my own.

To look at what I believe are our tasks are for this assignment, take a look at the "Group TODO List.txt" file that I've written up.